VA-Tech-Dev/requirements.txt
discord.py
flask
mysql-connector-python
flask-cors